rm .urs

+++ download_data_product.py
@@ -34,7 +34,8 @@ def download(source,file_name_gz,local):
     cmd = "curl -c .urs_cookies -b .urs_cookies -n -L {} -O".format(source+"/"+file_name_gz)
     try:
         result = subprocess.getstatusoutput(cmd)
-        os.remove(".urs_cookies")
+        if (os.path.exists(".urs_cookies")):
+            os.remove(".urs_cookies")
         if (os.path.exists(file_name_gz)):
             return True
         else:
